Handle wheel installs correctly (#16)
When installing MAX as a wheel, we don't ship a modular.cfg file we can use to launch CLI tools. This is mostly worked around on the packaging end, where CLI tools are wrapped to set the necessary environment variables directly, but we have to make some adjustments to the SDK loading code to work with the different structure.

There is one caveat: debugging the current Mojo file is not supported, because we can't launch the wrapped `mojo` executable under LLDB. We can handle this better in the future, but for now you just get an error.